+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python version `3.11` added to package classifiers
 - Python version `3.12` tested during CI
 - Python version `3.12` added to package classifiers
+- Support for `no_user_input` in `Acknowledge.display()` and `Echo.display()`.
 
 ### Fixed